youme/requirements.txt
python-dotenv==1.0.0
PyYAML==6.0.1
openai==1.52.0
pypdf2==3.0.1
pdf2image==1.17.0
fastapi==0.110.2
uvicorn==0.29.0
tiktoken==0.8.0
pandas==2.2.3
openpyxl==3.1.5